Add new text card

+++ src/components/Card.js
@@ -5,31 +5,39 @@ const { ipcRenderer } = require('electron');
 const fs = require('fs');
 const {basename} = require('path');
 
-const Card = ({ path, selected, onRemove }) => {
+const Card = ({ children, path, selected, onRemove, isFile }) => {
     const [isHovered, setIsHover] = useState(false);
     const [content, setContent] = useState(null);
     const [image, setImage] = useState(null);
 
     useEffect(() => {
-        fs.readFile(path, (error, data) => {
-            const bn = basename(path).split('.');
-            const ex = bn[bn.length - 1];
+        if(isFile) {
+            fs.readFile(path, (error, data) => {
+                const bn = basename(path).split('.');
+                const ex = bn[bn.length - 1];
 
-            if(/(png|jpg|jpeg|svg|gif)/.test(ex)) {
-                setImage(`data:image/${ex};base64,${data.toString('base64')}`);
-            } else if(/(txt)/.test(ex)) {
-                setContent(data.toString());
-            }
-        });
-    }, [path]);
+                if(/(png|jpg|jpeg|svg|gif)/.test(ex)) {
+                    setImage(`data:image/${ex};base64,${data.toString('base64')}`);
+                } else if(/(txt)/.test(ex)) {
+                    setContent(data.toString());
+                }
+            });
+        }
+    }, [path, isFile]);
 
-    const handleOnClickImage = useCallback(() => {
-        ipcRenderer.send('copy', path);
+    const handleOnClick = useCallback(() => {
+        if(isFile) {
+            ipcRenderer.send('copy', path);
+        } else {
+            ipcRenderer.send('new');
+        }
     }, [path]);
 
     const handleOnClickRemove = useCallback(() => {
-        ipcRenderer.send('remove', path);
-        onRemove(path);
+        if(isFile) {
+            ipcRenderer.send('remove', path);
+            onRemove(path);
+        }
     }, [path]);
 
     const handleHover = useCallback(value => {
@@ -41,13 +49,17 @@ const Card = ({ path, selected, onRemove }) => {
     return (
         <div onMouseOver={() => handleHover(true)} onMouseLeave={() => handleHover(false)} className={'pointer-events-none transition-all relative h-full w-full'}>
             <div className={'inline-block relative'}>
-                <div
-                    onClick={() => handleOnClickRemove()}
-                    className={`pointer-events-auto absolute cursor-pointer -right-4 top-0 p-2 transition-all rounded-full bg-background-accent hover:bg-background-hover opacity-${isHovered ? '100' : '0'} pointer-events-${isHovered ? 'auto' : 'none'}`}>
-                    <X color={'var(--color-text-default)'} />
-                </div>
+                {
+                    isFile && (
+                        <div
+                            onClick={() => handleOnClickRemove()}
+                            className={`pointer-events-auto absolute cursor-pointer -right-4 top-0 p-2 transition-all rounded-full bg-background-accent hover:bg-background-hover opacity-${isHovered ? '100' : '0'} pointer-events-${isHovered ? 'auto' : 'none'}`}>
+                            <X color={'var(--color-text-default)'} />
+                        </div>
+                    )
+                }
 
-                <div onClick={() => handleOnClickImage()}  className={`pointer-events-auto cursor-pointer w-card rounded-lg mt-5 border-2 transition-colors ${(isHovered || selected) ? 'border-primary-default' : 'border-transparent'} ${content ? 'p-8 bg-background-hover text-text-default' : ''}`}>
+                <div onClick={() => handleOnClick()}  className={`pointer-events-auto cursor-pointer w-card rounded-lg mt-5 border-2 transition-colors ${(isHovered || selected) ? 'border-primary-default' : 'border-transparent'} ${(content || !isFile) ? 'p-8 bg-background-hover text-text-default' : ''}`}>
                     {
                         image && (
                             <img
@@ -59,8 +71,8 @@ const Card = ({ path, selected, onRemove }) => {
                     }
 
                     {
-                        content && (
-                            <p className={'overflow-ellipsis overflow-hidden'}>{content}</p>
+                        (content || !isFile) && (
+                            <p className={'overflow-ellipsis overflow-hidden'}>{content || children}</p>
                         )
                     }
                 </div>
@@ -70,7 +82,8 @@ const Card = ({ path, selected, onRemove }) => {
 };
 
 Card.defaultProps = {
-    path: ''
+    path: '',
+    isFile: true
 };
 
 export default Card;
